consolodate ViewsHelpers files
