--- frontend/fingertips-frontend/components/views/ViewsHelpers.test.tsx
@@ -1,3 +1,0 @@
-describe('ViewsHelpers', () => {
-  it.todo('should make the correct API call - copy tests from View');
-});
